requirements.txt
Flask==3.0.2
pymongo==4.6.1
requests==2.31.0
python-dotenv==1.0.1
numpy>=1.26.0
gTTS==2.3.2
playsound==1.3.0
setuptools>=69.0.0
pip>=24.0.0
flask-cors==4.0.0
dnspython==2.6.1
schedule==1.2.1
google-generativeai==0.3.2
sounddevice==0.4.6
google-adk>=0.1.0
gunicorn==21.2.0
